requirements.txt
google-api-python-client==2.147.0
pandas==2.2.2
python-dateutil==2.9.0.post0
tzdata==2024.1
